+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.annotation.RequiredAdmin;
 import roomescape.auth.dto.LoginMember;
 import roomescape.auth.service.AuthService;
-import roomescape.reservation.dto.request.AdminReservationRequest;
-import roomescape.reservation.dto.request.ReservationCreateRequest;
-import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.service.dto.request.AdminReservationRequest;
+import roomescape.reservation.service.dto.request.ReservationCreateRequest;
+import roomescape.reservation.service.dto.response.ReservationResponse;
 import roomescape.reservation.service.CreateReservationService;
 
 import java.net.URI;
