+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -17,12 +17,12 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import roomescape.auth.dto.LoginMember;
-import roomescape.reservation.dto.response.BookedReservationTimeResponse;
-import roomescape.reservation.dto.request.FilteringReservationRequest;
-import roomescape.reservation.dto.response.MyReservationsResponse;
-import roomescape.reservation.dto.request.ReservationCreateRequest;
-import roomescape.reservation.dto.request.ReservationRequest;
-import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.service.dto.response.BookedReservationTimeResponse;
+import roomescape.reservation.service.dto.request.FilteringReservationRequest;
+import roomescape.reservation.service.dto.response.MyReservationsResponse;
+import roomescape.reservation.service.dto.request.ReservationCreateRequest;
+import roomescape.reservation.service.dto.request.ReservationRequest;
+import roomescape.reservation.service.dto.response.ReservationResponse;
 import roomescape.reservation.service.CreateReservationService;
 import roomescape.reservation.service.ReservationService;
