+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -5,11 +5,11 @@ import roomescape.auth.dto.LoginMember;
 import roomescape.common.exception.EntityNotFoundException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.dto.request.FilteringReservationRequest;
-import roomescape.reservation.dto.response.BookedReservationTimeResponse;
-import roomescape.reservation.dto.response.MyReservationsResponse;
-import roomescape.reservation.dto.response.ReservationResponse;
-import roomescape.reservation.dto.response.ReservationTimeResponse;
+import roomescape.reservation.service.dto.request.FilteringReservationRequest;
+import roomescape.reservation.service.dto.response.BookedReservationTimeResponse;
+import roomescape.reservation.service.dto.response.MyReservationsResponse;
+import roomescape.reservation.service.dto.response.ReservationResponse;
+import roomescape.reservation.service.dto.response.ReservationTimeResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.waiting.domain.Waiting;
